requirements.txt
google-api-python-client==2.154.0
streamlit==1.43.2
schedule==1.2.2
python-dotenv==1.0.1
requests==2.32.3
playwright==1.51.0
